requirements.txt
# Core dependencies
python-telegram-bot==21.0.1
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# AI & LangChain
langchain==0.1.0
langchain-community==0.0.13
langchain-groq==0.0.1

# Vector DB
chromadb==0.4.22
sentence-transformers==2.3.1

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23

# Utilities
requests==2.31.0
aiohttp==3.9.1
